+++ lib/src/server/server.dart
@@ -90,7 +90,9 @@ class ConnectionServer {
   final CodecRegistry? _codecRegistry;
   final GrpcErrorHandler? _errorHandler;
   final ServerKeepAliveOptions _keepAliveOptions;
-  final Map<ServerTransportConnection, List<ServerHandler>> _handlers = {};
+
+  @visibleForTesting
+  final Map<ServerTransportConnection, List<ServerHandler>> handlers = {};
 
   final _connections = <ServerTransportConnection>[];
 
@@ -117,7 +119,7 @@ class ConnectionServer {
     InternetAddress? remoteAddress,
   }) async {
     _connections.add(connection);
-    _handlers[connection] = [];
+    handlers[connection] = [];
     // TODO(jakobr): Set active state handlers, close connection after idle
     // timeout.
     final onDataReceivedController = StreamController<void>();
@@ -129,12 +131,14 @@ class ConnectionServer {
       dataNotifier: onDataReceivedController.stream,
     ).handle();
     connection.incomingStreams.listen((stream) {
-      _handlers[connection]!.add(serveStream_(
+      final handler = serveStream_(
         stream: stream,
         clientCertificate: clientCertificate,
         remoteAddress: remoteAddress,
         onDataReceived: onDataReceivedController.sink,
-      ));
+      );
+      handler.onCanceled.then((_) => handlers[connection]?.remove(handler));
+      handlers[connection]!.add(handler);
     }, onError: (error, stackTrace) {
       if (error is Error) {
         Zone.current.handleUncaughtError(error, stackTrace);
@@ -144,11 +148,11 @@ class ConnectionServer {
       // half-closed tcp streams.
       // Half-closed  streams seems to not be fully supported by package:http2.
       // https://github.com/dart-lang/http2/issues/42
-      for (var handler in _handlers[connection]!) {
+      for (var handler in handlers[connection]!) {
         handler.cancel();
       }
       _connections.remove(connection);
-      _handlers.remove(connection);
+      handlers.remove(connection);
       await onDataReceivedController.close();
     });
   }